Refactor data utilities into modules
