+++ src/data/mapping.ts
@@ -1,0 +1,27 @@
+import type { SupportedTypesWithMapped } from "../types";
+
+const typeMappings = new Map<string, keyof SupportedTypesWithMapped>([
+  ["AMMO", "item"],
+  ["GUN", "item"],
+  ["ARMOR", "item"],
+  ["PET_ARMOR", "item"],
+  ["TOOL", "item"],
+  ["TOOLMOD", "item"],
+  ["TOOL_ARMOR", "item"],
+  ["BOOK", "item"],
+  ["COMESTIBLE", "item"],
+  ["CONTAINER", "item"],
+  ["ENGINE", "item"],
+  ["WHEEL", "item"],
+  ["GUNMOD", "item"],
+  ["MAGAZINE", "item"],
+  ["BATTERY", "item"],
+  ["GENERIC", "item"],
+  ["BIONIC_ITEM", "item"],
+  ["MONSTER", "monster"],
+  ["city_building", "overmap_special"],
+]);
+
+export const mapType = (
+  type: keyof SupportedTypesWithMapped,
+): keyof SupportedTypesWithMapped => typeMappings.get(type) ?? type;
